update pages 는 페이지 import 만 하도록
